chore: refactor `TaskManager` API to allow integration of external TMs
